update use hooks for spilt func
